WIP: Solo rpc
